Add a file for every View
The structs which where previously on the same folder
were now moved to their own folder
